+++ orix/quaternion/orientation.py
@@ -31,6 +31,7 @@ import numpy as np
 from scipy.spatial.transform import Rotation as SciPyRotation
 from tqdm import tqdm
 
+from orix._util import deprecated
 from orix.quaternion.orientation_region import OrientationRegion
 from orix.quaternion.rotation import Rotation
 from orix.quaternion.symmetry import C1, Symmetry, _get_unique_symmetry_elements
@@ -142,6 +143,7 @@ class Misorientation(Rotation):
         equivalent = Gr.outer(orientations.outer(Gl))
         return self.__class__(equivalent).flatten()
 
+    @deprecated(since="0.12", removal="0.13", alternative="reduce")
     def map_into_symmetry_reduced_zone(self, verbose: bool = False) -> Misorientation:
         """Return equivalent transformations which have the smallest
         angle of rotation as a new misorientation.
@@ -164,25 +166,87 @@ class Misorientation(Rotation):
         >>> m.symmetry = (C4, C2)
         >>> m.map_into_symmetry_reduced_zone()
         Misorientation (2,) 4, 2
-        [[-0.7071  0.7071  0.      0.    ]
+        [[-0.7071 0.     -0.7071  0.    ]
         [ 0.      1.      0.      0.    ]]
         """
-        Gl, Gr = self._symmetry
-        symmetry_pairs = iproduct(Gl, Gr)
+        # Combine symmetry elements of start and end of transformation
+        # given by the (mis)orientation
+        start, end = self._symmetry
+        symmetry_pairs = iproduct(start, end)
         if verbose:
-            symmetry_pairs = tqdm(symmetry_pairs, total=Gl.size * Gr.size)
+            symmetry_pairs = tqdm(symmetry_pairs, total=start.size * end.size)
 
-        orientation_region = OrientationRegion.from_symmetry(Gl, Gr)
-        o_inside = self.__class__.identity(self.shape)
-        outside = np.ones(self.shape, dtype=bool)
-        for gl, gr in symmetry_pairs:
-            o_transformed = gl * self[outside] * gr
-            o_inside[outside] = o_transformed
-            outside = ~(o_inside < orientation_region)
-            if not np.any(outside):
+        # Find the (mis)orientations which lie inside the Rodrigues
+        # (orientation) or MacKenzie (misorientation) fundamental zone
+        # (FZ), given by the symmetry elements. We loop over all
+        # symmetry pairs and rotate all (mis)orientations until all are
+        # inside the FZ.
+        fz = OrientationRegion.from_symmetry(start, end)
+        reduced = self.__class__.identity(self.shape)
+        is_outside = np.ones(self.shape, dtype=bool)
+        for sym_start, sym_end in symmetry_pairs:
+            rotated = sym_end * self[is_outside] * sym_start
+            reduced[is_outside] = rotated
+            is_outside = ~(reduced < fz)
+            if not is_outside.any():
                 break
-        o_inside._symmetry = (Gl, Gr)
-        return o_inside
+        reduced._symmetry = (start, end)
+
+        return reduced
+
+    def reduce(self) -> Misorientation:
+        """Return the reduced misorientations with the smallest
+        rotation angles.
+
+        Returns
+        -------
+        reduced
+            Reduced misorientations.
+
+        Notes
+        -----
+        The misorientation with the smallest rotation angle is the one
+        inside the misorientation fundamental zone (FZ, asymmetric
+        domain) given by the proper point groups of :attr:`symmetry`.
+        The definition of the FZ follows the procedure in section 5.3.1
+        in :cite:`martineau2020multivariate`.
+
+        An alternative description of finding the Rodrigues FZ is given
+        in :cite:`morawiec1996rodrigues`, which is the basis for
+        reduction of (mis)orientations in EMsoft.
+
+        Examples
+        --------
+        >>> from orix.quaternion.symmetry import C4, C2
+        >>> data = np.array([[0.5, 0.5, 0.5, 0.5], [0, 1, 0, 0]])
+        >>> m = Misorientation(data)
+        >>> m.symmetry = (C4, C2)
+        >>> m.map_into_symmetry_reduced_zone()
+        Misorientation (2,) 4, 2
+        [[-0.7071 0.     -0.7071  0.    ]
+        [ 0.      1.      0.      0.    ]]
+        """
+        # Combine symmetry elements of start and end of transformation
+        # given by the misorientations
+        start, end = self._symmetry
+        symmetry_pairs = iproduct(start, end)
+
+        # Find the misorientations which lie inside the
+        # misorientation fundamental zone (FZ), given by the symmetry
+        # elements. We loop over all symmetry pairs and rotate
+        # misorientations until all are inside the FZ.
+        fz = OrientationRegion.from_symmetry(start, end)
+        reduced = self.__class__.identity(self.shape)
+        is_outside = np.ones(self.shape, dtype=bool)
+        for sym_start, sym_end in symmetry_pairs:
+            rotated = sym_end * self[is_outside] * sym_start
+            reduced[is_outside] = rotated
+            is_outside = ~(reduced < fz)
+            if not is_outside.any():
+                break
+        reduced._symmetry = (start, end)
+
+        return reduced
 
     def __repr__(self):
         """String representation."""
@@ -264,20 +328,13 @@ class Misorientation(Rotation):
         # Plot wireframe
         if wireframe_kwargs is None:
             wireframe_kwargs = {}
-        if isinstance(self.symmetry, tuple):
-            fundamental_zone = OrientationRegion.from_symmetry(
-                s1=self.symmetry[0], s2=self.symmetry[1]
-            )
-            ax.plot_wireframe(fundamental_zone, **wireframe_kwargs)
-        else:
-            # Orientation via inheritance
-            fundamental_zone = OrientationRegion.from_symmetry(self.symmetry)
-            ax.plot_wireframe(fundamental_zone, **wireframe_kwargs)
+        fz = OrientationRegion.from_symmetry(*self._symmetry)
+        ax.plot_wireframe(fz, **wireframe_kwargs)
 
         # Correct the aspect ratio of the axes according to the extent
         # of the boundaries of the fundamental region, and also restrict
         # the data limits to these boundaries
-        ax._correct_aspect_ratio(fundamental_zone)
+        ax._correct_aspect_ratio(fz)
 
         ax.axis("off")
         figure.subplots_adjust(left=0, right=1, bottom=0, top=1, hspace=0, wspace=0)
@@ -286,7 +343,7 @@ class Misorientation(Rotation):
             to_plot = self.get_random_sample(size)
         else:
             to_plot = self
-        ax.scatter(to_plot, fundamental_zone=fundamental_zone, **kwargs)
+        ax.scatter(to_plot, fundamental_zone=fz, **kwargs)
 
         if return_figure:
             return figure
@@ -594,10 +651,10 @@ class Orientation(Misorientation):
             # Call to Object3d.squeeze() doesn't carry over symmetry
             misorientation = Misorientation(self * ~other).squeeze()
             misorientation.symmetry = (self.symmetry, other.symmetry)
-            return misorientation.map_into_symmetry_reduced_zone()
+            return misorientation.reduce()
         return NotImplemented
 
-    # TODO: Remove use of **kwargs in 1.0
+    # TODO: Remove use of **kwargs in 0.13
     @classmethod
     def from_euler(
         cls,
@@ -1208,6 +1265,29 @@ class Orientation(Misorientation):
         )
 
         return euler_in_region
+
+    def reduce(self) -> Orientation:
+        """Return the reduced orientations with the smallest rotation
+        angles.
+
+        Returns
+        -------
+        reduced
+            Reduced orientations.
+
+        Notes
+        -----
+        The orientation with the smallest rotation angle is the one
+        inside the Rodrigues fundamental zone (FZ, asymmetric domain)
+        given by the proper point group of :attr:`symmetry`. The
+        definition of the FZ follows the procedure in section 5.3.1 in
+        :cite:`martineau2020multivariate`.
+
+        An alternative description of finding the Rodrigues FZ is given
+        in :cite:`morawiec1996rodrigues`, which is the basis for
+        reduction of orientations in EMsoft.
+        """
+        return super().reduce()
 
     def scatter(
         self,