+++ orix/quaternion/orientation_region.py
@@ -16,40 +16,22 @@
 # You should have received a copy of the GNU General Public License
 # along with orix.  If not, see <http://www.gnu.org/licenses/>.
 
-"""An orientation region is some subset of the complete space of orientations.
-
-The complete orientation space represents every possible orientation of an
-object. The whole space is not always needed, for example if the orientation
-of an object is constrained or (most commonly) if the object is symmetrical. In
-this case, the space can be segmented using sets of Rotations representing
-boundaries in the space. This is clearest in the Rodrigues parametrisation,
-where the boundaries are planes, such as the example here: the asymmetric
-domain of an adjusted 432 symmetry.
-
-.. image:: /_static/img/orientation-region-Oq.png
-   :width: 300px
-   :alt: Boundaries of an orientation region in Rodrigues space.
-   :align: center
-
-Rotations or orientations can be inside or outside of an orientation region.
-"""
-
 from __future__ import annotations
 
 import itertools
-from typing import Tuple
+from typing import List, Tuple
 
 import numpy as np
 
 from orix.quaternion import Quaternion
 from orix.quaternion.rotation import Rotation
 from orix.quaternion.symmetry import C1, Symmetry, get_distinguished_points
-from orix.vector import AxAngle, Rodrigues
+from orix.vector import Rodrigues
 
 _EPSILON = 1e-9  # small number to avoid round off problems
 
 
-def _get_large_cell_normals(s1, s2):
+def _get_large_cell_normals(s1: Symmetry, s2: Symmetry) -> Rotation:
     dp = get_distinguished_points(s1, s2)
     normals = Rodrigues.zero(dp.shape + (2,))
     planes1 = dp.axis * np.tan(dp.angle / 4)
@@ -57,9 +39,7 @@ def _get_large_cell_normals(s1, s2):
     planes2.data[np.isnan(planes2.data)] = 0
     normals[:, 0] = planes1
     normals[:, 1] = planes2
-    normals: Rotation = (
-        Rotation.from_neo_euler(normals).flatten().unique(antipodal=False)
-    )
+    normals = Rotation.from_neo_euler(normals).flatten().unique(antipodal=False)
     if not normals.size:
         return normals
     _, inv = normals.axis.unique(return_inverse=True)
@@ -68,10 +48,8 @@ def _get_large_cell_normals(s1, s2):
     for i in np.unique(inv):
         n = normals[inv == i]
         axes_unique.append(n.axis.data[0])
-        angles_unique.append(n.angle.max())
-    normals = Rotation.from_neo_euler(
-        AxAngle.from_axes_angles(np.array(axes_unique), angles_unique)
-    )
+        angles_unique.append(np.max(n.angle))
+    normals = Rotation.from_axes_angles(np.array(axes_unique), angles_unique)
     return normals
 
 
@@ -117,47 +95,73 @@ def get_proper_groups(Gl: Symmetry, Gr: Symmetry) -> Tuple[Symmetry, Symmetry]:
             return Gl.laue_proper_subgroup, Gr.proper_subgroup
         else:
             raise NotImplementedError(
-                "Both groups are improper, " "and do not contain inversion."
+                "Both groups are improper and do not contain inversion"
             )
 
 
 class OrientationRegion(Rotation):
     """A set of :class:`~orix.quaternion.Rotation` which are the normals
     of an orientation region.
+
+    Notes
+    -----
+    An orientation region is some subset of the complete space of
+    orientations.
+
+    The complete orientation space represents every possible orientation
+    of an object. The whole space is not always needed, for example if
+    the orientation of an object is constrained or (most commonly) if
+    the object is symmetrical. In this case, the space can be segmented
+    using sets of Rotations representing boundaries in the space. This
+    is clearest in the Rodrigues parametrisation, where the boundaries
+    are planes, such as the example here: the asymmetric domain of an
+    adjusted 432 symmetry.
+
+    .. image:: /_static/img/orientation-region-Oq.png
+       :width: 300px
+       :alt: Boundaries of an orientation region in Rodrigues space.
+       :align: center
+
+    Rotations or orientations can be inside or outside of an orientation
+    region.
     """
 
     @classmethod
     def from_symmetry(cls, s1: Symmetry, s2: Symmetry = C1) -> OrientationRegion:
-        """The set of unique (mis)orientations of a symmetrical object.
+        """Return the set of unique (mis)orientations of a symmetrical
+        object.
 
         Parameters
         ----------
         s1
             First symmetry.
         s2
-            Second symmetry.
+            Second symmetry. Default is C1 (the identity).
+
+        Returns
+        -------
+        region
+            Orientation region.
         """
         s1, s2 = get_proper_groups(s1, s2)
         large_cell_normals = _get_large_cell_normals(s1, s2)
         disjoint = s1 & s2
         fz = disjoint.fundamental_zone()
-        fz_normals = Rotation.from_neo_euler(AxAngle.from_axes_angles(fz, np.pi))
+        fz_normals = Rotation.from_axes_angles(fz, np.pi)
         normals = Rotation(np.concatenate([large_cell_normals.data, fz_normals.data]))
-        orientation_region = cls(normals)
-        vertices = orientation_region.vertices()
-        if vertices.size:
-            orientation_region = orientation_region[
-                np.any(np.isclose(orientation_region.dot_outer(vertices), 0), axis=1)
-            ]
-        return orientation_region
+        region = cls(normals)
+        vertices = region.vertices()
+        if vertices.size > 0:
+            region = region[np.any(np.isclose(region.dot_outer(vertices), 0), axis=1)]
+        return region
 
     def vertices(self) -> Rotation:
-        """Return the vertices of the asymmetric domain.
+        """Return the vertices of the orientation region.
 
         Returns
         -------
         rot
-            Domain vertices.
+            Region vertices.
         """
         normal_combinations = list(itertools.combinations(self, 3))
         if len(normal_combinations) < 1:
@@ -174,7 +178,15 @@ class OrientationRegion(Rotation):
         surface = np.any(np.isclose(rot.dot_outer(self), 0), axis=1)
         return rot[surface]
 
-    def faces(self) -> list:
+    def faces(self) -> List[Rotation]:
+        """Return the faces of the orientation region.
+
+        Returns
+        -------
+        faces
+            List of sets of rotations, each set describing a face or the
+            region.
+        """
         normals = Rotation(self)
         vertices = self.vertices()
         faces = []
@@ -183,10 +195,10 @@ class OrientationRegion(Rotation):
         faces = [f for f in faces if f.size > 2]
         return faces
 
-    def __gt__(self, other: OrientationRegion) -> np.ndarray:
+    def __gt__(self, other: Quaternion) -> np.ndarray:
         """Overridden greater than method. Applying this to an
         Orientation will return only those orientations that lie within
-        the OrientationRegion.
+        the orientation region.
         """
         c = Quaternion(self).dot_outer(Quaternion(other))
         inside = np.logical_or(
@@ -196,21 +208,28 @@ class OrientationRegion(Rotation):
         return inside
 
     def get_plot_data(self) -> Rotation:
-        """Suitable Rotations for the construction of a wireframe."""
+        """Return suitable rotations for the construction of a
+        wireframe delineating the borders of the region.
+
+        Returns
+        -------
+        g
+            Rotations delineating the borders of the region.
+        """
         from orix.vector import Vector3d
 
         # Get a grid of vector directions
         theta = np.linspace(0, 2 * np.pi - _EPSILON, 361)
         rho = np.linspace(0, np.pi - _EPSILON, 181)
         theta, rho = np.meshgrid(theta, rho)
-        g = Vector3d.from_polar(rho, theta)
+        v = Vector3d.from_polar(rho, theta)
 
-        # Get the cell vector normal norms
+        # Get the norms of the cell vector normals
         n = Rodrigues.from_rotation(self).norm[:, np.newaxis, np.newaxis]
         if n.size == 0:
-            return Rotation.from_neo_euler(AxAngle.from_axes_angles(g, np.pi))
+            return Rotation.from_axes_angles(v, np.pi)
 
-        d = (-self.axis).dot_outer(g.unit)
+        d = (-self.axis).dot_outer(v.unit)
         x = n * d
         with np.errstate(divide="ignore"):
             omega = 2 * np.arctan(np.where(x != 0, x**-1, np.pi))
@@ -218,6 +237,6 @@ class OrientationRegion(Rotation):
         # Keep the smallest allowed angle
         omega[omega < 0] = np.pi
         omega = np.min(omega, axis=0)
-        r = Rotation.from_neo_euler(AxAngle.from_axes_angles(g.unit, omega))
+        g = Rotation.from_axes_angles(v.unit, omega)
 
-        return r
+        return g