+++ CHANGELOG.rst
@@ -11,18 +11,30 @@ Unreleased
 
 Added
 -----
+- ``Vector3d.get_path()`` class method to get vectors between two or more vectors.
 
 Changed
 -------
+- The ``reduce()`` method (replacing ``map_into_symmetry_reduced_zone()``) for
+  the ``Misorientation`` and ``Orientation`` classes now apply symmetry operators in the
+  opposite order of what was done previously.
+- The ``convention`` parameter in ``from_euler()`` and ``to_euler()`` will be removed in
+  the next minor release, 0.13, instead of release 1.0 as previously stated.
 
 Deprecated
 ----------
+- ``map_into_symmetry_reduced_zone()`` is deprecated since 0.12 and will be removed in
+  0.13. Use ``reduce()`` instead.
 
 Removed
 -------
+- ``verbose`` parameter in ``reduce()``
+  (replacing ``map_into_symmetry_reduced_zone()``).
 
 Fixed
 -----
+- Transparency of polar stereographic grid lines can now be controlled by Matplotlib's
+  ``grid.alpha``, just like the azimuth grid lines.
 
 2023-03-14 - version 0.11.1
 ===========================